fix(AL8-RULE-003): split Unix socket adapter
